requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://docs.astronomer.io/astro/runtime-image-architecture#provider-packages
apache-airflow-providers-amazon==8.11.0
apache-airflow-providers-common-sql==1.8.0
apache-airflow-providers-snowflake==5.1.1  # the Snowflake provider is needed to create a Snowflake connection!